Add a vinz-shim greasemonkey script
This script can be placed at the /vinz endpoint of
review.openstack.org(this is a 404).
Once there it will load a real webpage from localhost(from the
perspective of the browser) and replace the on review.o.o with the
provided html. This gives us a real application to develop but still has
the cookies and headers to authenticate to gerrit.